Anova Test

Provides a pipe-friendly framework to performs ANOVA tests. Wrapper around the function Anova() in the car package.

anova_test(data, formula, ...)

Arguments

data

a data.frame containing the variables in the formula.

formula

a formula of the form x ~ group where x is a numeric variable giving the data values and group is a factor with one or multiple levels giving the corresponding groups. For example, formula = TP53 ~ cancer_group.

...

other arguments to be passed to the function Anova.

Value

return a data frame with the following columns:

.y.: the y variable used in the test.

term: the grouping variable.

statistic: Test statistic (F-statistic) used to compute the p-value.

p: p-value.

method: the statistical test used to compare groups.

Examples

# Load data
#:::::::::::::::::::::::::::::::::::::::
data("ToothGrowth")
df <- ToothGrowth

# One-way Anova test
#:::::::::::::::::::::::::::::::::::::::::
df %>% anova_test(len ~ dose)

#> # A tibble: 1 x 4
#>     .y.  term statistic       p
#>   <chr> <chr>     <dbl>   <dbl>
#> 1   len  dose  105.0648 1.2e-14

# Two-way Anova test
#:::::::::::::::::::::::::::::::::::::::::
df %>% anova_test(len ~ supp*dose)

#> # A tibble: 3 x 4
#>     .y.      term  statistic       p
#>   <chr>     <chr>      <dbl>   <dbl>
#> 1   len      supp  12.317020 8.9e-04
#> 2   len      dose 133.415146 1.9e-16
#> 3   len supp:dose   5.333483 2.5e-02