Box's M-test for Homogeneity of Covariance Matrices

Performs the Box's M-test for homogeneity of covariance matrices obtained from multivariate normal data according to one grouping variable. The test is based on the chi-square approximation.

box_m(data, group)

Arguments

data

a numeric data.frame or matrix containing n observations of p variables; it is expected that n > p.

group

a vector of length n containing the class of each observation; it is usually a factor.

Value

A data frame containing the following components:

statistic an approximated value of the chi-square distribution.

parameter the degrees of freedom related of the test statistic in this case that it follows a Chi-square distribution.

p.value the p-value of the test.

method the character string "Box's M-test for Homogeneity of Covariance Matrices".

Examples

data(iris)
box_m(iris[, -5], iris[, 5])
#> # A tibble: 1 × 4
#>   statistic  p.value parameter method                                           
#>       <dbl>    <dbl>     <dbl> <chr>                                            
#> 1      141. 3.35e-20        20 Box's M-test for Homogeneity of Covariance Matri…