box_m {rstatix}	R Documentation

Box's M-test for Homogeneity of Covariance Matrices

Description

Performs the Box's M-test for homogeneity of covariance matrices obtained from multivariate normal data according to one grouping variable. The test is based on the chi-square approximation.

Usage

box_m(data, group)

Arguments

data
a numeric data.frame or matrix containing n observations of p variables; it is expected that n > p.

group
a vector of length n containing the class of each observation; it is usually a factor.

Value

A data frame containing the following components:

statistic an approximated value of the chi-square distribution.

parameter the degrees of freedom related of the test statistic in this case that it follows a Chi-square distribution.

p.value the p-value of the test.

method the character string "Box's M-test for Homogeneity of Covariance Matrices".

Examples

data(iris)
box_m(iris[, -5], iris[, 5])

[Package rstatix version 0.7.0 Index]