Correlation Test

Provides a pipe-friendly framework to perform correlation test between paired samples, using Pearson, Kendall or Spearman method. Wrapper around the function cor.test().

Can also performs multiple pairwise correlation analyses between more than two variables or between two different vectors of variables. Using this function, you can also compute, for example, the correlation between one variable vs many.

cor_test(data, ..., vars = NULL, vars2 = NULL,
  alternative = "two.sided", method = "pearson", conf.level = 0.95,
  use = "pairwise.complete.obs")

Arguments

data
a data.frame containing the variables.

...
One or more unquoted expressions (or variable names) separated by commas. Used to select a variable of interest. Alternative to the argument vars.

vars
optional character vector containing variable names for correlation analysis. Ignored when dot vars are specified.

If vars is NULL, multiple pairwise correlation tests is performed between all variables in the data.

If vars contain only one variable, a pairwise correlation analysis is performed between the specified variable vs either all the remaining numeric variables in the data or variables in vars2 (if specified).

If vars contain two or more variables: i) if vars2 is not specified, a pairwise correlation analysis is performed between all possible combinations of variables. ii) if vars2 is specified, each element in vars is tested against all elements in vars2

. Accept unquoted variable names: c(var1, var2).

vars2
optional character vector. If specified, each element in vars is tested against all elements in vars2. Accept unquoted variable names: c(var1, var2).

alternative
indicates the alternative hypothesis and must be one of "two.sided", "greater" or "less". You can specify just the initial letter. "greater" corresponds to positive association, "less" to negative association.

method
a character string indicating which correlation coefficient is to be used for the test. One of "pearson", "kendall", or "spearman", can be abbreviated.

conf.level
confidence level for the returned confidence interval. Currently only used for the Pearson product moment correlation coefficient if there are at least 4 complete pairs of observations.

use
an optional character string giving a method for computing covariances in the presence of missing values. This must be (an abbreviation of) one of the strings "everything", "all.obs", "complete.obs", "na.or.complete", or "pairwise.complete.obs".

Value

return a data frame with the following columns:

var1, var2: the variables used in the correlation test.

cor: the correlation coefficient.

statistic: Test statistic used to compute the p-value.

p: p-value.

conf.low,conf.high: Lower and upper bounds on a confidence interval.

method: the method used to compute the statistic.

Functions

cor_test: correlation test between two or more variables.

See also

Examples

# Correlation between the specified variable vs
# the remaining numeric variables in the data
#:::::::::::::::::::::::::::::::::::::::::
mtcars %>% cor_test(mpg)

#> [90m# A tibble: 10 x 8[39m
#>    var1  var2    cor statistic        p conf.low conf.high method 
#>  [90m*[39m [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<chr>[39m[23m  
#> [90m 1[39m mpg   cyl   -[31m0[39m[31m.[39m[31m85[39m     -[31m8[39m[31m.[39m[31m92[39m 6.11[90me[39m[31m-10[39m  -[31m0[39m[31m.[39m[31m926[39m     -[31m0[39m[31m.[39m[31m716[39m Pearson
#> [90m 2[39m mpg   disp  -[31m0[39m[31m.[39m[31m85[39m     -[31m8[39m[31m.[39m[31m75[39m 9.38[90me[39m[31m-10[39m  -[31m0[39m[31m.[39m[31m923[39m     -[31m0[39m[31m.[39m[31m708[39m Pearson
#> [90m 3[39m mpg   hp    -[31m0[39m[31m.[39m[31m78[39m     -[31m6[39m[31m.[39m[31m74[39m 1.79[90me[39m[31m- 7[39m  -[31m0[39m[31m.[39m[31m885[39m     -[31m0[39m[31m.[39m[31m586[39m Pearson
#> [90m 4[39m mpg   drat   0.68      5.10 1.78[90me[39m[31m- 5[39m   0.436      0.832 Pearson
#> [90m 5[39m mpg   wt    -[31m0[39m[31m.[39m[31m87[39m     -[31m9[39m[31m.[39m[31m56[39m 1.29[90me[39m[31m-10[39m  -[31m0[39m[31m.[39m[31m934[39m     -[31m0[39m[31m.[39m[31m744[39m Pearson
#> [90m 6[39m mpg   qsec   0.42      2.53 1.71[90me[39m[31m- 2[39m   0.082[4m0[24m     0.670 Pearson
#> [90m 7[39m mpg   vs     0.66      4.86 3.42[90me[39m[31m- 5[39m   0.410      0.822 Pearson
#> [90m 8[39m mpg   am     0.6       4.11 2.85[90me[39m[31m- 4[39m   0.318      0.784 Pearson
#> [90m 9[39m mpg   gear   0.48      3.00 5.40[90me[39m[31m- 3[39m   0.158      0.710 Pearson
#> [90m10[39m mpg   carb  -[31m0[39m[31m.[39m[31m55[39m     -[31m3[39m[31m.[39m[31m62[39m 1.08[90me[39m[31m- 3[39m  -[31m0[39m[31m.[39m[31m755[39m     -[31m0[39m[31m.[39m[31m250[39m Pearson

# Correlation test between two variables
#:::::::::::::::::::::::::::::::::::::::::
mtcars %>% cor_test(wt, mpg)

#> [90m# A tibble: 1 x 8[39m
#>   var1  var2    cor statistic        p conf.low conf.high method 
#> [90m*[39m [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<chr>[39m[23m  
#> [90m1[39m wt    mpg   -[31m0[39m[31m.[39m[31m87[39m     -[31m9[39m[31m.[39m[31m56[39m 1.29[90me[39m[31m-10[39m   -[31m0[39m[31m.[39m[31m934[39m    -[31m0[39m[31m.[39m[31m744[39m Pearson

# Pairwise correlation between multiple variables
#:::::::::::::::::::::::::::::::::::::::::
mtcars %>% cor_test(wt, mpg, disp)

#> [90m# A tibble: 9 x 8[39m
#>   var1  var2    cor     statistic         p conf.low conf.high method 
#> [90m*[39m [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m [3m[90m<dbl>[39m[23m         [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<chr>[39m[23m  
#> [90m1[39m wt    wt     1     367[4m5[24m[4m7[24m[4m0[24m386.   2.27[90me[39m[31m-236[39m    1.000     1     Pearson
#> [90m2[39m wt    mpg   -[31m0[39m[31m.[39m[31m87[39m         -[31m9[39m[31m.[39m[31m56[39m 1.29[90me[39m[31m- 10[39m   -[31m0[39m[31m.[39m[31m934[39m    -[31m0[39m[31m.[39m[31m744[39m Pearson
#> [90m3[39m wt    disp   0.89         10.6  1.22[90me[39m[31m- 11[39m    0.781     0.944 Pearson
#> [90m4[39m mpg   wt    -[31m0[39m[31m.[39m[31m87[39m         -[31m9[39m[31m.[39m[31m56[39m 1.29[90me[39m[31m- 10[39m   -[31m0[39m[31m.[39m[31m934[39m    -[31m0[39m[31m.[39m[31m744[39m Pearson
#> [90m5[39m mpg   mpg    1           [31mInf[39m    0.  [90m [39m        1         1     Pearson
#> [90m6[39m mpg   disp  -[31m0[39m[31m.[39m[31m85[39m         -[31m8[39m[31m.[39m[31m75[39m 9.38[90me[39m[31m- 10[39m   -[31m0[39m[31m.[39m[31m923[39m    -[31m0[39m[31m.[39m[31m708[39m Pearson
#> [90m7[39m disp  wt     0.89         10.6  1.22[90me[39m[31m- 11[39m    0.781     0.944 Pearson
#> [90m8[39m disp  mpg   -[31m0[39m[31m.[39m[31m85[39m         -[31m8[39m[31m.[39m[31m75[39m 9.38[90me[39m[31m- 10[39m   -[31m0[39m[31m.[39m[31m923[39m    -[31m0[39m[31m.[39m[31m708[39m Pearson
#> [90m9[39m disp  disp   1           [31mInf[39m    0.  [90m [39m        1         1     Pearson

# Grouped data
#:::::::::::::::::::::::::::::::::::::::::
iris %>%
  group_by(Species) %>%
 cor_test(Sepal.Width, Sepal.Length)

#> [90m# A tibble: 3 x 9[39m
#>   Species   var1     var2       cor statistic        p conf.low conf.high method
#> [90m*[39m [3m[90m<fct>[39m[23m     [3m[90m<chr>[39m[23m    [3m[90m<chr>[39m[23m    [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<chr>[39m[23m 
#> [90m1[39m setosa    Sepal.W… Sepal.L…  0.74      7.68 6.71[90me[39m[31m-10[39m    0.585     0.846 Pears…
#> [90m2[39m versicol… Sepal.W… Sepal.L…  0.53      4.28 8.77[90me[39m[31m- 5[39m    0.290     0.702 Pears…
#> [90m3[39m virginica Sepal.W… Sepal.L…  0.46      3.56 8.43[90me[39m[31m- 4[39m    0.205     0.653 Pears…

# Multiple correlation test
#:::::::::::::::::::::::::::::::::::::::::
# Correlation between one variable vs many
mtcars %>% cor_test(
  vars = "mpg",
  vars2 = c("disp", "hp", "drat")
 )

#> [90m# A tibble: 3 x 8[39m
#>   var1  var2    cor statistic        p conf.low conf.high method 
#> [90m*[39m [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<chr>[39m[23m  
#> [90m1[39m mpg   disp  -[31m0[39m[31m.[39m[31m85[39m     -[31m8[39m[31m.[39m[31m75[39m 9.38[90me[39m[31m-10[39m   -[31m0[39m[31m.[39m[31m923[39m    -[31m0[39m[31m.[39m[31m708[39m Pearson
#> [90m2[39m mpg   hp    -[31m0[39m[31m.[39m[31m78[39m     -[31m6[39m[31m.[39m[31m74[39m 1.79[90me[39m[31m- 7[39m   -[31m0[39m[31m.[39m[31m885[39m    -[31m0[39m[31m.[39m[31m586[39m Pearson
#> [90m3[39m mpg   drat   0.68      5.10 1.78[90me[39m[31m- 5[39m    0.436     0.832 Pearson

# Correlation between two vectors of variables
# Each element in vars is tested against all elements in vars2
mtcars %>% cor_test(
  vars = c("mpg", "wt"),
  vars2 = c("disp", "hp", "drat")
 )

#> [90m# A tibble: 6 x 8[39m
#>   var1  var2    cor statistic        p conf.low conf.high method 
#> [90m*[39m [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m [3m[90m<chr>[39m[23m  
#> [90m1[39m mpg   disp  -[31m0[39m[31m.[39m[31m85[39m     -[31m8[39m[31m.[39m[31m75[39m 9.38[90me[39m[31m-10[39m   -[31m0[39m[31m.[39m[31m923[39m    -[31m0[39m[31m.[39m[31m708[39m Pearson
#> [90m2[39m mpg   hp    -[31m0[39m[31m.[39m[31m78[39m     -[31m6[39m[31m.[39m[31m74[39m 1.79[90me[39m[31m- 7[39m   -[31m0[39m[31m.[39m[31m885[39m    -[31m0[39m[31m.[39m[31m586[39m Pearson
#> [90m3[39m mpg   drat   0.68      5.10 1.78[90me[39m[31m- 5[39m    0.436     0.832 Pearson
#> [90m4[39m wt    disp   0.89     10.6  1.22[90me[39m[31m-11[39m    0.781     0.944 Pearson
#> [90m5[39m wt    hp     0.66      4.80 4.15[90me[39m[31m- 5[39m    0.403     0.819 Pearson
#> [90m6[39m wt    drat  -[31m0[39m[31m.[39m[31m71[39m     -[31m5[39m[31m.[39m[31m56[39m 4.78[90me[39m[31m- 6[39m   -[31m0[39m[31m.[39m[31m850[39m    -[31m0[39m[31m.[39m[31m484[39m Pearson