Correlation Test

Provides a pipe-friendly framework to perform correlation test between paired samples, using Pearson, Kendall or Spearman method. Wrapper around the function cor.test().

Available functions include:

cor_test(): Correlation test between two variables.

mcor_test(): Multiple correlation tests between two vectors of variables.

cor_test(data, formula, alternative = "two.sided", method = "pearson",
  conf.level = 0.95, use = "complete.obs", ...)

mcor_test(data, x, y, ...)

Arguments

data

a data.frame containing the variables in the formula.

formula

a formula of the form var1 ~ var2 where var1 and var2 are two numeric variables in the data.

alternative

indicates the alternative hypothesis and must be one of "two.sided", "greater" or "less". You can specify just the initial letter. "greater" corresponds to positive association, "less" to negative association.

method

a character string indicating which correlation coefficient is to be used for the test. One of "pearson", "kendall", or "spearman", can be abbreviated.

conf.level

confidence level for the returned confidence interval. Currently only used for the Pearson product moment correlation coefficient if there are at least 4 complete pairs of observations.

use

an optional character string giving a method for computing covariances in the presence of missing values. This must be (an abbreviation of) one of the strings "everything", "all.obs", "complete.obs", "na.or.complete", or "pairwise.complete.obs".

...

other arguments to be passed to the function cor.test.

x, y

character vectors containing variable names to be used in the correlation analysis.

Value

return a data frame with the following columns:

var1, var2: the variables used in the correlation test.

cor: the correlation coefficient.

statistic: Test statistic used to compute the p-value.

p: p-value.

conf.low,conf.high: Lower and upper bounds on a confidende interval.

method: the method used to compute the statistic.

Functions

cor_test: Correlation test between two variables.

mcor_test: Multiple correlation tests between two vectors of variables.

Examples

# Correlation test between two variables
#:::::::::::::::::::::::::::::::::::::::::
mtcars %>% cor_test(wt ~ mpg)

#> # A tibble: 1 x 8
#>   var1  var2    cor statistic        p conf.low conf.high method 
#>   <chr> <chr> <dbl>     <dbl>    <dbl>    <dbl>     <dbl> <chr>  
#> 1 wt    mpg   -0.87     -9.56 1.29e-10   -0.934    -0.744 Pearson

# Grouped data
#:::::::::::::::::::::::::::::::::::::::::
iris %>%
  group_by(Species) %>%
 cor_test(Sepal.Width ~ Sepal.Length)

#> # A tibble: 3 x 9
#>   Species   var1     var2      cor statistic        p conf.low conf.high method
#>   <fct>     <chr>    <chr>   <dbl>     <dbl>    <dbl>    <dbl>     <dbl> <chr> 
#> 1 setosa    Sepal.W… Sepal.…  0.74      7.68 6.71e-10    0.585     0.846 Pears…
#> 2 versicol… Sepal.W… Sepal.…  0.53      4.28 8.77e- 5    0.290     0.702 Pears…
#> 3 virginica Sepal.W… Sepal.…  0.46      3.56 8.43e- 4    0.205     0.653 Pears…

# Multiple correlation test
#:::::::::::::::::::::::::::::::::::::::::
# Correlation between one variable vs many
mtcars %>% mcor_test(x = "mpg", y = c("disp", "hp", "drat"))

#> # A tibble: 3 x 8
#>   var1  var2    cor statistic        p conf.low conf.high method 
#>   <chr> <chr> <dbl>     <dbl>    <dbl>    <dbl>     <dbl> <chr>  
#> 1 mpg   disp  -0.85     -8.75 9.38e-10   -0.923    -0.708 Pearson
#> 2 mpg   hp    -0.78     -6.74 1.79e- 7   -0.885    -0.586 Pearson
#> 3 mpg   drat   0.68      5.10 1.78e- 5    0.436     0.832 Pearson

# Correlation between two vectors of variables
mtcars %>% mcor_test(x = c("mpg", "wt"), y = c("disp", "hp", "drat"))

#> # A tibble: 6 x 8
#>   var1  var2    cor statistic        p conf.low conf.high method 
#>   <chr> <chr> <dbl>     <dbl>    <dbl>    <dbl>     <dbl> <chr>  
#> 1 mpg   disp  -0.85     -8.75 9.38e-10   -0.923    -0.708 Pearson
#> 2 mpg   hp    -0.78     -6.74 1.79e- 7   -0.885    -0.586 Pearson
#> 3 mpg   drat   0.68      5.10 1.78e- 5    0.436     0.832 Pearson
#> 4 wt    disp   0.89     10.6  1.22e-11    0.781     0.944 Pearson
#> 5 wt    hp     0.66      4.80 4.15e- 5    0.403     0.819 Pearson
#> 6 wt    drat  -0.71     -5.56 4.78e- 6   -0.850    -0.484 Pearson