Compute Cramer's V

Compute Cramer's V, which measures the strength of the association between categorical variables.

cramer_v(x, y = NULL, correct = TRUE, ...)

Arguments

x

a numeric vector or matrix. x and y can also both be factors.

y

a numeric vector; ignored if x is a matrix. If x is a factor, y should be a factor of the same length.

correct

a logical indicating whether to apply continuity correction when computing the test statistic for 2 by 2 tables: one half is subtracted from all \(|O - E|\) differences; however, the correction will not be bigger than the differences themselves. No correction is done if simulate.p.value = TRUE.

...

other arguments passed to the function chisq.test().

Examples

# Data preparation
df <- as.table(rbind(c(762, 327, 468), c(484, 239, 477)))
dimnames(df) <- list(
  gender = c("F", "M"),
  party = c("Democrat","Independent", "Republican")
)
df
#>       party
#> gender Democrat Independent Republican
#>      F      762         327        468
#>      M      484         239        477
# Compute cramer's V
cramer_v(df)
#> [1] 0.1044358