df_group_by {rstatix}	R Documentation

Group a Data Frame by One or more Variables

Description

Group a data frame by one or more variables. Supports standard and non standard evaluation.

Usage

df_group_by(data, ..., vars = NULL)

Arguments

data
a data frame

...
One or more unquoted expressions (or variable names) separated by commas. Used to select a variable of interest.

vars
a character vector containing the variable names of interest.

Examples

# Non standard evaluation
by_dose <- head(ToothGrowth) %>%
   df_group_by(dose)
by_dose

# Standard evaluation
head(ToothGrowth) %>%
   df_group_by(vars = c("dose", "supp"))

[Package rstatix version 0.7.2 Index]