Alternative to dplyr::do for Doing Anything

Provides a flexible alternative to the dplyr:do() function. Technically it uses nest() + mutate() + map() to apply arbitrary computation to a grouped data frame.

The output is a data frame. If the applied function returns a data frame, then the output will be automatically unnested. Otherwise, the output includes the grouping variables and a column named ".results." (by default), which is a "list-columns" containing the results for group combinations.

doo(data, .f, ..., result = ".results.")

Arguments

data
a (grouped) data frame

.f
A function, formula, or atomic vector. For example ~t.test(len ~ supp, data = .).

...
Additional arguments passed on to .f

result
the column name to hold the results. Default is ".results.".

Value

a data frame

Examples

# Example 1: pipe-friendly t_test().
# Two possibilities of usage are available
#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
# Use this
ToothGrowth %>%
  group_by(dose) %>%
  doo(~t_test(data =., len ~ supp))

#> [90m# A tibble: 3 x 7[39m
#>    dose .y.   group1 group2 statistic    df       p
#>   [3m[90m<dbl>[39m[23m [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m      [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m   [3m[90m<dbl>[39m[23m
#> [90m1[39m   0.5 len   OJ     VC        3.17    15.0 0.006[4m3[24m[4m6[24m
#> [90m2[39m   1   len   OJ     VC        4.03    15.4 0.001[4m0[24m[4m4[24m
#> [90m3[39m   2   len   OJ     VC       -[31m0[39m[31m.[39m[31m0[39m[31m46[4m1[24m[39m  14.0 0.964  

# Or this
ToothGrowth %>%
  group_by(dose) %>%
  doo(t_test, len ~ supp)

#> [90m# A tibble: 3 x 7[39m
#>    dose .y.   group1 group2 statistic    df       p
#>   [3m[90m<dbl>[39m[23m [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m      [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m   [3m[90m<dbl>[39m[23m
#> [90m1[39m   0.5 len   OJ     VC        3.17    15.0 0.006[4m3[24m[4m6[24m
#> [90m2[39m   1   len   OJ     VC        4.03    15.4 0.001[4m0[24m[4m4[24m
#> [90m3[39m   2   len   OJ     VC       -[31m0[39m[31m.[39m[31m0[39m[31m46[4m1[24m[39m  14.0 0.964  

# Example 2: R base function t.test() (not pipe friendly)
# One possibility of usage
#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
comparisons <- ToothGrowth %>%
   group_by(dose) %>%
   doo(~t.test(len ~ supp, data =.))
comparisons

#> [90m# A tibble: 3 x 2[39m
#>    dose .results.  
#>   [3m[90m<dbl>[39m[23m [3m[90m<list>[39m[23m     
#> [90m1[39m   0.5 [90m<S3: htest>[39m
#> [90m2[39m   1   [90m<S3: htest>[39m
#> [90m3[39m   2   [90m<S3: htest>[39m

comparisons$.results.

#> [[1]]
#> 
#> 	Welch Two Sample t-test
#> 
#> data:  len by supp
#> t = 3.1697, df = 14.969, p-value = 0.006359
#> alternative hypothesis: true difference in means is not equal to 0
#> 95 percent confidence interval:
#>  1.719057 8.780943
#> sample estimates:
#> mean in group OJ mean in group VC 
#>            13.23             7.98 
#> 
#> 
#> [[2]]
#> 
#> 	Welch Two Sample t-test
#> 
#> data:  len by supp
#> t = 4.0328, df = 15.358, p-value = 0.001038
#> alternative hypothesis: true difference in means is not equal to 0
#> 95 percent confidence interval:
#>  2.802148 9.057852
#> sample estimates:
#> mean in group OJ mean in group VC 
#>            22.70            16.77 
#> 
#> 
#> [[3]]
#> 
#> 	Welch Two Sample t-test
#> 
#> data:  len by supp
#> t = -0.046136, df = 14.04, p-value = 0.9639
#> alternative hypothesis: true difference in means is not equal to 0
#> 95 percent confidence interval:
#>  -3.79807  3.63807
#> sample estimates:
#> mean in group OJ mean in group VC 
#>            26.06            26.14 
#> 
#> 

# Example 3: R base function combined with tidy()
#%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
ToothGrowth %>%
   group_by(dose) %>%
   doo(~t.test(len ~ supp, data =.) %>% tidy())

#> [90m# A tibble: 3 x 11[39m
#>    dose estimate estimate1 estimate2 statistic p.value parameter conf.low
#>   [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m   [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m
#> [90m1[39m   0.5     5.25      13.2      7.98    3.17   0.006[4m3[24m[4m6[24m      15.0     1.72
#> [90m2[39m   1       5.93      22.7     16.8     4.03   0.001[4m0[24m[4m4[24m      15.4     2.80
#> [90m3[39m   2      -[31m0[39m[31m.[39m[31m0[39m[31m8[39m      26.1     26.1    -[31m0[39m[31m.[39m[31m0[39m[31m46[4m1[24m[39m 0.964        14.0    -[31m3[39m[31m.[39m[31m80[39m
#> [90m# … with 3 more variables: conf.high [3m[90m<dbl>[90m[23m, method [3m[90m<chr>[90m[23m, alternative [3m[90m<chr>[90m[23m[39m