Compute Summary Statistics

Compute summary statistics for one or multiple numeric variables.

get_summary_stats(data, ..., type = c("full", "common", "robust",
  "five_number", "mean_sd", "mean_se", "mean_ci", "median_iqr",
  "median_mad"), show = NULL)

Arguments

data
a data frame

...
(optional) One or more unquoted expressions (or variable names) separated by commas. Used to select a variable of interest. If no variable is specified, then the summary statistics of all numeric variables in the data frame is computed.

type
type of summary statistics. Possible values include: "full", "common", "robust", "five_number", "mean_sd", "mean_se", "mean_ci", "median_iqr", "median_mad"

show
a character vector specifying the summary statistics you want to show. Example: show = c("n", "mean", "sd"). This is used to filter the output after computation.

Value

A data frame containing descriptive statistics, such as:

n: the number of individuals

min: minimum

max: maximum

median: median

mean: mean

q1, q3: the first and the third quartile, respectively.

iqr: interquartile range

mad: median absolute deviation (see ?MAD)

sd: standard deviation of the mean

se: standard error of the mean

ci: 95 percent confidence interval of the mean

Examples

# Full summary statistics
data("ToothGrowth")
ToothGrowth %>% get_summary_stats(len)

#> [90m# A tibble: 1 x 13[39m
#>   variable     n   min   max median    q1    q3   iqr   mad  mean    sd    se
#>   [3m[90m<chr>[39m[23m    [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m  [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m
#> [90m1[39m len         60   4.2  33.9   19.2  13.1  25.3  12.2  9.04  18.8  7.65 0.988
#> [90m# … with 1 more variable: ci [3m[90m<dbl>[90m[23m[39m

# Summary statistics of grouped data
# Show only common summary
ToothGrowth %>%
  group_by(dose, supp) %>%
  get_summary_stats(len, type = "common")

#> [90m# A tibble: 6 x 12[39m
#>    dose supp  variable     n   min   max median   iqr  mean    sd    se    ci
#>   [3m[90m<dbl>[39m[23m [3m[90m<fct>[39m[23m [3m[90m<chr>[39m[23m    [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m  [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m
#> [90m1[39m   0.5 OJ    len         10   8.2  21.5  12.2   6.48 13.2   4.46 1.41   3.19
#> [90m2[39m   0.5 VC    len         10   4.2  11.5   7.15  4.95  7.98  2.75 0.869  1.96
#> [90m3[39m   1   OJ    len         10  14.5  27.3  23.4   5.35 22.7   3.91 1.24   2.80
#> [90m4[39m   1   VC    len         10  13.6  22.5  16.5   2.02 16.8   2.52 0.795  1.80
#> [90m5[39m   2   OJ    len         10  22.4  30.9  26.0   2.5  26.1   2.66 0.84   1.90
#> [90m6[39m   2   VC    len         10  18.5  33.9  26.0   5.42 26.1   4.80 1.52   3.43

# Robust summary statistics
ToothGrowth %>% get_summary_stats(len, type = "robust")

#> [90m# A tibble: 1 x 4[39m
#>   variable     n median   iqr
#>   [3m[90m<chr>[39m[23m    [3m[90m<dbl>[39m[23m  [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m
#> [90m1[39m len         60   19.2  12.2

# Five number summary statistics
ToothGrowth %>% get_summary_stats(len, type = "five_number")

#> [90m# A tibble: 1 x 7[39m
#>   variable     n   min   max    q1 median    q3
#>   [3m[90m<chr>[39m[23m    [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m  [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m
#> [90m1[39m len         60   4.2  33.9  13.1   19.2  25.3

# Compute only mean and sd
ToothGrowth %>% get_summary_stats(len, type = "mean_sd")

#> [90m# A tibble: 1 x 4[39m
#>   variable     n  mean    sd
#>   [3m[90m<chr>[39m[23m    [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m
#> [90m1[39m len         60  18.8  7.65

# Compute full summary statistics but show only mean, sd, median, iqr
ToothGrowth %>%
    get_summary_stats(len, show = c("mean", "sd", "median", "iqr"))

#> [90m# A tibble: 1 x 6[39m
#>   variable     n  mean    sd median   iqr
#>   [3m[90m<chr>[39m[23m    [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m  [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m
#> [90m1[39m len         60  18.8  7.65   19.2  12.2