Kruskal-Wallis Test

Provides a pipe-friendly framework to performs Kruskal-Wallis rank sum test. Wrapper around the function kruskal.test().

kruskal_test(data, formula, ...)

Arguments

data
a data.frame containing the variables in the formula.

formula
a formula of the form x ~ group where x is a numeric variable giving the data values and group is a factor with one or multiple levels giving the corresponding groups. For example, formula = TP53 ~ cancer_group.

...
other arguments to be passed to the function kruskal.test.

Value

return a data frame with the following columns:

.y.: the y variable used in the test.

statistic: the kruskal-wallis rank sum statistic used to compute the p-value.

p: p-value.

method: the statistical test used to compare groups.

Examples

# Load data
#:::::::::::::::::::::::::::::::::::::::
data("ToothGrowth")
df <- ToothGrowth

# Kruskal-wallis rank sum test
#:::::::::::::::::::::::::::::::::::::::::
df %>% kruskal_test(len ~ dose)

#> # A tibble: 1 x 4
#>   .y.   statistic            p method        
#>   <chr>     <dbl>        <dbl> <chr>         
#> 1 len        40.7 0.0000000015 Kruskal-Wallis