Shapiro-Wilk Normality Test

Provides a pipe-friendly framework to performs Shapiro-Wilk test of normality. Support grouped data and multiple variables for multivariate normality tests. Wrapper around the R base function shapiro.test(). Can handle grouped data.

shapiro_test(data, ..., vars = NULL)

mshapiro_test(data)

Arguments

data
a data frame. Columns are variables.

...
One or more unquoted expressions (or variable names) separated by commas. Used to select a variable of interest.

vars
optional character vector containing variable names.Ignored when dot vars are specified.

Value

a data frame containing the value of the Shapiro-Wilk statistic and the corresponding p.value.

Functions

shapiro_test: univariate Shapiro-Wilk normality test

mshapiro_test: multivariate Shapiro-Wilk normality test. This is a modified copy of the mshapiro.test() function of the package mvnormtest, for internal convenience.

Examples

# Shapiro Wilk normality test for one variable
iris %>% shapiro_test(Sepal.Length)

#> [90m# A tibble: 1 x 3[39m
#>   variable     statistic      p
#>   [3m[90m<chr>[39m[23m            [3m[90m<dbl>[39m[23m  [3m[90m<dbl>[39m[23m
#> [90m1[39m Sepal.Length     0.976 0.010[4m2[24m

# Shapiro Wilk normality test for two variables
iris %>% shapiro_test(Sepal.Length, Petal.Width)

#> [90m# A tibble: 2 x 3[39m
#>   variable     statistic            p
#>   [3m[90m<chr>[39m[23m            [3m[90m<dbl>[39m[23m        [3m[90m<dbl>[39m[23m
#> [90m1[39m Petal.Width      0.902 0.000[4m0[24m[4m0[24m[4m0[24m016[4m8[24m
#> [90m2[39m Sepal.Length     0.976 0.010[4m2[24m      

# Multivariate normality test
mshapiro_test(iris[, 1:3])

#> [90m# A tibble: 1 x 2[39m
#>   statistic p.value
#>       [3m[90m<dbl>[39m[23m   [3m[90m<dbl>[39m[23m
#> [90m1[39m     0.991   0.443