shapiro_test {rstatix}	R Documentation

Shapiro-Wilk Normality Test

Description

Provides a pipe-friendly framework to performs Shapiro-Wilk test of normality. Support grouped data and multiple variables for multivariate normality tests. Wrapper around the R base function shapiro.test(). Can handle grouped data. Read more: Normality Test in R.

Usage

shapiro_test(data, ..., vars = NULL)

mshapiro_test(data)

Arguments

data
a data frame. Columns are variables.

...
One or more unquoted expressions (or variable names) separated by commas. Used to select a variable of interest.

vars
optional character vector containing variable names. Ignored when dot vars are specified.

Value

a data frame containing the value of the Shapiro-Wilk statistic and the corresponding p.value.

Functions

shapiro_test: univariate Shapiro-Wilk normality test

mshapiro_test: multivariate Shapiro-Wilk normality test. This is a modified copy of the mshapiro.test() function of the package mvnormtest, for internal convenience.

Examples

# Shapiro Wilk normality test for one variable
iris %>% shapiro_test(Sepal.Length)

# Shapiro Wilk normality test for two variables
iris %>% shapiro_test(Sepal.Length, Petal.Width)

# Multivariate normality test
mshapiro_test(iris[, 1:3])

[Package rstatix version 0.7.0 Index]