T-test

Provides a pipe-friendly framework to performs one and two sample t-tests.

t_test(data, formula, comparisons = NULL, ref.group = NULL,
  p.adjust.method = "holm", paired = FALSE, var.equal = FALSE,
  alternative = "two.sided", mu = 0, conf.level = 0.95)

one_sample_t_test(data, formula, mu = 0, ...)

two_sample_t_test(data, formula, paired = FALSE, ...)

pairwise_t_test(data, formula, comparisons = NULL, ref.group = NULL,
  p.adjust.method = "holm", ...)

pairwise_t_test_psd(data, formula, comparisons = NULL,
  ref.group = NULL, p.adjust.method = "holm", pool.sd = TRUE,
  alternative = "two.sided")

one_vs_all_t_test(data, formula, p.adjust.method = "holm", ...)

Arguments

data
a data.frame containing the variables in the formula.

formula
a formula of the form x ~ group where x is a numeric variable giving the data values and group is a factor with one or multiple levels giving the corresponding groups. For example, formula = TP53 ~ cancer_group.

comparisons
A list of length-2 vectors specifying the groups of interest to be compared. For example to compare groups "A" vs "B" and "B" vs "C", the argument is as follow: comparisons = list(c("A", "B"), c("B", "C"))

ref.group
a character string specifying the reference group. If specified, for a given grouping variable, each of the group levels will be compared to the reference group (i.e. control group).

p.adjust.method
method to adjust p values for multiple comparisons. Used when pairwise comparisons are performed. Allowed values include "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". If you don't want to adjust the p value (not recommended), use p.adjust.method = "none".

paired
a logical indicating whether you want a paired test.

var.equal
a logical variable indicating whether to treat the two variances as being equal. If TRUE then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used.

alternative
a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.

mu
a number indicating the true value of the mean (or difference in means if you are performing a two sample test).

conf.level
confidence level of the interval.

...
other arguments to be passed to the function t.test.

pool.sd
switch to allow/disallow the use of a pooled SD.

Value

return a data frame with the following columns:

.y.: the y variable used in the test.

group1,group2: the compared groups in the pairwise tests.

statistic: Test statistic used to compute the p-value.

p: p-value.

p.adj: the adjusted p-value.

method: the statistical test used to compare groups.

p.signif, p.adj.signif: the significance level of p-values and adjusted p-values, respectively.

Details

- pairwise_t_test() applies the standard two sample t-test to all possible pairs of groups. This method calls the t.test(), so extra arguments are accepted.

- pairwise_t_test_psd() performs a pairwise t-test with pooled SD. It is a wrapper around the R base function pairwise.t.test(), which calculates, by default, a common SD for all groups and uses that for all comparisons (this can be useful if some groups are small). This method does not actually call t.test, so extra arguments are ignored. Pooling does not generalize to paired tests so pool.sd and paired cannot both be TRUE.

- If a list of comparisons is specified, the result of the pairwise tests is filtered to keep only the comparisons of interest.The p-value is adjusted after filtering.

- For a grouped data, if pairwise test is performed, then the p-values are adjusted for each group level independently.

Functions

one_sample_t_test: performs one-sample t-test.

two_sample_t_test: performs two sample t-test.

pairwise_t_test: performs pairwise two sample t-test.

pairwise_t_test_psd: performs pairwise t-test with pooled SD. Wrapper around the R base function pairwise.t.test.

one_vs_all_t_test: performs pairwise two sample t-test comparing each grouping variable levels against all (i.e. basemean)

Examples

# Load data
#:::::::::::::::::::::::::::::::::::::::
data("ToothGrowth")
df <- ToothGrowth

# One-sample test
#:::::::::::::::::::::::::::::::::::::::::
df %>% t_test(len ~ 1, mu = 0)

#> # A tibble: 1 x 6
#>   .y.   group1 group2     statistic        p method
#>   <chr> <chr>  <chr>          <dbl>    <dbl> <chr> 
#> 1 len   1      null model      19.1 6.90e-27 T-test

# Two-samples unpaired test
#:::::::::::::::::::::::::::::::::::::::::
df %>% t_test(len ~ supp)

#> # A tibble: 1 x 6
#>   .y.   group1 group2 statistic     p method
#>   <chr> <chr>  <chr>      <dbl> <dbl> <chr> 
#> 1 len   OJ     VC          1.92 0.061 T-test

# Two-samples paired test
#:::::::::::::::::::::::::::::::::::::::::
df %>% t_test (len ~ supp, paired = TRUE)

#> # A tibble: 1 x 6
#>   .y.   group1 group2 statistic      p method
#>   <chr> <chr>  <chr>      <dbl>  <dbl> <chr> 
#> 1 len   OJ     VC          3.30 0.0025 T-test

# Compare supp levels after grouping the data by "dose"
#::::::::::::::::::::::::::::::::::::::::
df %>%
  group_by(dose) %>%
  t_test(data =., len ~ supp) %>%
  adjust_pvalue() %>%
  add_significance("p.adj")

#> # A tibble: 3 x 9
#>    dose .y.   group1 group2 statistic      p method  p.adj p.adj.signif
#>   <dbl> <chr> <chr>  <chr>      <dbl>  <dbl> <chr>   <dbl> <chr>       
#> 1   0.5 len   OJ     VC        3.17   0.0064 T-test 0.0128 *           
#> 2   1   len   OJ     VC        4.03   0.001  T-test 0.003  **          
#> 3   2   len   OJ     VC       -0.0461 0.96   T-test 0.96   ns          

# pairwise comparisons
#::::::::::::::::::::::::::::::::::::::::
# As dose contains more thant two levels ==>
# pairwise test is automatically performed.
df %>% t_test(len ~ dose)

#> # A tibble: 3 x 9
#>   .y.   group1 group2 statistic        p method    p.adj p.signif p.adj.signif
#>   <chr> <chr>  <chr>      <dbl>    <dbl> <chr>     <dbl> <chr>    <chr>       
#> 1 len   0.5    1          -6.48 1.30e- 7 T-test 2.60e- 7 ****     ****        
#> 2 len   0.5    2         -11.8  4.40e-14 T-test 1.30e-13 ****     ****        
#> 3 len   1      2          -4.90 1.90e- 5 T-test 1.90e- 5 ****     ****        

# Comparison against reference group
#::::::::::::::::::::::::::::::::::::::::
# each level is compared to the ref group
df %>% t_test(len ~ dose, ref.group = "0.5")

#> # A tibble: 2 x 9
#>   .y.   group1 group2 statistic        p method    p.adj p.signif p.adj.signif
#>   <chr> <chr>  <chr>      <dbl>    <dbl> <chr>     <dbl> <chr>    <chr>       
#> 1 len   0.5    1          -6.48 1.30e- 7 T-test 1.30e- 7 ****     ****        
#> 2 len   0.5    2         -11.8  4.40e-14 T-test 8.80e-14 ****     ****        

# Comparison against all
#::::::::::::::::::::::::::::::::::::::::
df %>% t_test(len ~ dose, ref.group = "all")

#> # A tibble: 3 x 9
#>   .y.   group1 group2 statistic          p method    p.adj p.signif p.adj.signif
#>   <chr> <chr>  <chr>      <dbl>      <dbl> <chr>     <dbl> <chr>    <chr>       
#> 1 len   all    0.5        5.82     2.90e-7 T-test  8.70e-7 ****     ****        
#> 2 len   all    1         -0.660    5.10e-1 T-test  5.10e-1 ns       ns          
#> 3 len   all    2         -5.61     4.30e-7 T-test  8.70e-7 ****     ****