T-test

Provides a pipe-friendly framework to performs one and two sample t-tests.

t_test(data, formula, comparisons = NULL, ref.group = NULL,
  p.adjust.method = "holm", paired = FALSE, var.equal = FALSE,
  alternative = "two.sided", mu = 0, conf.level = 0.95,
  detailed = FALSE)

pairwise_t_test(data, formula, comparisons = NULL, ref.group = NULL,
  p.adjust.method = "holm", paired = FALSE, pool.sd = !paired,
  detailed = FALSE, ...)

Arguments

data
a data.frame containing the variables in the formula.

formula
a formula of the form x ~ group where x is a numeric variable giving the data values and group is a factor with one or multiple levels giving the corresponding groups. For example, formula = TP53 ~ cancer_group.

comparisons
A list of length-2 vectors specifying the groups of interest to be compared. For example to compare groups "A" vs "B" and "B" vs "C", the argument is as follow: comparisons = list(c("A", "B"), c("B", "C"))

ref.group
a character string specifying the reference group. If specified, for a given grouping variable, each of the group levels will be compared to the reference group (i.e. control group).

If ref.group = "all", pairwise two sample t-tests are performed for comparing each grouping variable levels against all (i.e. basemean).

p.adjust.method
method to adjust p values for multiple comparisons. Used when pairwise comparisons are performed. Allowed values include "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". If you don't want to adjust the p value (not recommended), use p.adjust.method = "none".

paired
a logical indicating whether you want a paired test.

var.equal
a logical variable indicating whether to treat the two variances as being equal. If TRUE then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used.

alternative
a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.

mu
a number indicating the true value of the mean (or difference in means if you are performing a two sample test).

conf.level
confidence level of the interval.

detailed
logical value. Default is FALSE. If TRUE, a detailed result is shown.

pool.sd
logical value used in the function pairwise_t_test(). Switch to allow/disallow the use of a pooled SD.

The pool.sd = TRUE (default) calculates a common SD for all groups and uses that for all comparisons (this can be useful if some groups are small). This method does not actually call t.test, so extra arguments are ignored. Pooling does not generalize to paired tests so pool.sd and paired cannot both be TRUE.

If pool.sd = FALSE the standard two sample t-test is applied to all possible pairs of groups. This method calls the t.test(), so extra arguments, such as var.equal are accepted.

...
other arguments to be passed to the function t.test.

Value

return a data frame with some the following columns:

.y.: the y variable used in the test.

group1,group2: the compared groups in the pairwise tests.

statistic: Test statistic used to compute the p-value.

df: degrees of freedom.

p: p-value.

p.adj: the adjusted p-value.

method: the statistical test used to compare groups.

p.signif, p.adj.signif: the significance level of p-values and adjusted p-values, respectively.

estimate: estimate of the effect size. It corresponds to the estimated mean or difference in means depending on whether it was a one-sample test or a two-sample test.

estimate1, estimate2: show the mean values of the two groups, respectively, for independent samples t-tests.

alternative: a character string describing the alternative hypothesis.

conf.low,conf.high: Lower and upper bound on a confidence interval.

Details

- If a list of comparisons is specified, the result of the pairwise tests is filtered to keep only the comparisons of interest. The p-value is adjusted after filtering.

- For a grouped data, if pairwise test is performed, then the p-values are adjusted for each group level independently.

Functions

t_test: t test

pairwise_t_test: performs pairwise two sample t-test. Wrapper around the R base function pairwise.t.test.

Examples

# Load data
#:::::::::::::::::::::::::::::::::::::::
data("ToothGrowth")
df <- ToothGrowth

# One-sample test
#:::::::::::::::::::::::::::::::::::::::::
df %>% t_test(len ~ 1, mu = 0)

#> [90m# A tibble: 1 x 6[39m
#>   .y.   group1 group2     statistic    df        p
#>   [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m          [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m
#> [90m1[39m len   1      null model      19.1    59 6.94[90me[39m[31m-27[39m

# Two-samples unpaired test
#:::::::::::::::::::::::::::::::::::::::::
df %>% t_test(len ~ supp)

#> [90m# A tibble: 1 x 6[39m
#>   .y.   group1 group2 statistic    df      p
#>   [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m      [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m  [3m[90m<dbl>[39m[23m
#> [90m1[39m len   OJ     VC          1.92  55.3 0.060[4m6[24m

# Two-samples paired test
#:::::::::::::::::::::::::::::::::::::::::
df %>% t_test (len ~ supp, paired = TRUE)

#> [90m# A tibble: 1 x 6[39m
#>   .y.   group1 group2 statistic    df       p
#>   [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m      [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m   [3m[90m<dbl>[39m[23m
#> [90m1[39m len   OJ     VC          3.30    29 0.002[4m5[24m[4m5[24m

# Compare supp levels after grouping the data by "dose"
#::::::::::::::::::::::::::::::::::::::::
df %>%
  group_by(dose) %>%
  t_test(data =., len ~ supp) %>%
  adjust_pvalue(method = "bonferroni") %>%
  add_significance("p.adj")

#> [90m# A tibble: 3 x 8[39m
#>   .y.   group1 group2 statistic    df       p   p.adj p.adj.signif
#>   [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m      [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m   [3m[90m<dbl>[39m[23m   [3m[90m<dbl>[39m[23m [3m[90m<chr>[39m[23m       
#> [90m1[39m len   OJ     VC        3.17    15.0 0.006[4m3[24m[4m6[24m 0.019[4m1[24m  *           
#> [90m2[39m len   OJ     VC        4.03    15.4 0.001[4m0[24m[4m4[24m 0.003[4m1[24m[4m2[24m **          
#> [90m3[39m len   OJ     VC       -[31m0[39m[31m.[39m[31m0[39m[31m46[4m1[24m[39m  14.0 0.964   1       ns          

# pairwise comparisons
#::::::::::::::::::::::::::::::::::::::::
# As dose contains more than two levels ==>
# pairwise test is automatically performed.
df %>% t_test(len ~ dose)

#> [90m# A tibble: 3 x 8[39m
#>   .y.   group1 group2 statistic    df        p    p.adj p.adj.signif
#>   [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m      [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m [3m[90m<chr>[39m[23m       
#> [90m1[39m len   0.5    1          -[31m6[39m[31m.[39m[31m48[39m  38.0 1.27[90me[39m[31m- 7[39m 2.54[90me[39m[31m- 7[39m ****        
#> [90m2[39m len   0.5    2         -[31m11[39m[31m.[39m[31m8[39m   36.9 4.40[90me[39m[31m-14[39m 1.32[90me[39m[31m-13[39m ****        
#> [90m3[39m len   1      2          -[31m4[39m[31m.[39m[31m90[39m  37.1 1.91[90me[39m[31m- 5[39m 1.91[90me[39m[31m- 5[39m ****        

# Comparison against reference group
#::::::::::::::::::::::::::::::::::::::::
# each level is compared to the ref group
df %>% t_test(len ~ dose, ref.group = "0.5")

#> [90m# A tibble: 2 x 8[39m
#>   .y.   group1 group2 statistic    df        p    p.adj p.adj.signif
#>   [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m      [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m [3m[90m<chr>[39m[23m       
#> [90m1[39m len   0.5    1          -[31m6[39m[31m.[39m[31m48[39m  38.0 1.27[90me[39m[31m- 7[39m 1.27[90me[39m[31m- 7[39m ****        
#> [90m2[39m len   0.5    2         -[31m11[39m[31m.[39m[31m8[39m   36.9 4.40[90me[39m[31m-14[39m 8.80[90me[39m[31m-14[39m ****        

# Comparison against all
#::::::::::::::::::::::::::::::::::::::::
df %>% t_test(len ~ dose, ref.group = "all")

#> [90m# A tibble: 3 x 8[39m
#>   .y.   group1 group2 statistic    df           p      p.adj p.adj.signif
#>   [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m      [3m[90m<dbl>[39m[23m [3m[90m<dbl>[39m[23m       [3m[90m<dbl>[39m[23m      [3m[90m<dbl>[39m[23m [3m[90m<chr>[39m[23m       
#> [90m1[39m len   all    0.5        5.82   56.4 0.000[4m0[24m[4m0[24m[4m0[24m290 0.000[4m0[24m[4m0[24m[4m0[24m87 ****        
#> [90m2[39m len   all    1         -[31m0[39m[31m.[39m[31m660[39m  57.5 0.512       0.512      ns          
#> [90m3[39m len   all    2         -[31m5[39m[31m.[39m[31m61[39m   66.5 0.000[4m0[24m[4m0[24m[4m0[24m425 0.000[4m0[24m[4m0[24m[4m0[24m87 ****