Tukey Honest Significant Differences

Provides a pipe-friendly framework to performs Tukey post-hoc tests. Wrapper around the function TukeyHSD().

Can handle different inputs formats: aov, lm, formula.

tukey_hsd(x, ...)

# S3 method for default
tukey_hsd(x, ...)

# S3 method for lm
tukey_hsd(x, ...)

# S3 method for data.frame
tukey_hsd(x, formula, ...)

Arguments

x
an object of class aov, lm or data.frame containing the variables used in the formula.

...
other arguments passed to the function TukeyHSD(). These include:

which: A character vector listing terms in the fitted model for which the intervals should be calculated. Defaults to all the terms.

ordered: A logical value indicating if the levels of the factor should be ordered according to increasing average in the sample before taking differences. If ordered is true then the calculated differences in the means will all be positive. The significant differences will be those for which the lwr end point is positive.

formula
a formula of the form x ~ group where x is a numeric variable giving the data values and group is a factor with one or multiple levels giving the corresponding groups. For example, formula = TP53 ~ cancer_group.

data
a data.frame containing the variables in the formula.

Value

a tibble data frame containing the results of the different comparisons.

Methods (by class)

default: performs tukey post-hoc test from aov() results.

lm: performs tukey post-hoc test from lm() model.

data.frame: performs tukey post-hoc tests using data and formula as inputs. ANOVA will be automatically performed using the function aov()

Examples

# Data preparation
df <- ToothGrowth
df$dose <- as.factor(df$dose)
# Tukey HSD from ANOVA results
aov(len ~ dose, data = df) %>% tukey_hsd()

#> [90m# A tibble: 3 x 8[39m
#>   term  group1 group2 comparison estimate conf.low conf.high    p.adj
#>   [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m         [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m
#> [90m1[39m dose  0.5    1      1-0.5          9.13     5.90     12.4  2.00[90me[39m[31m- 8[39m
#> [90m2[39m dose  0.5    2      2-0.5         15.5     12.3      18.7  1.12[90me[39m[31m-11[39m
#> [90m3[39m dose  1      2      2-1            6.36     3.14      9.59 4.25[90me[39m[31m- 5[39m

# two-way anova with interaction
aov(len ~ dose*supp, data = df) %>% tukey_hsd()

#> [90m# A tibble: 19 x 8[39m
#>    term      group1 group2 comparison    estimate conf.low conf.high    p.adj
#>    [3m[90m<chr>[39m[23m     [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m            [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m
#> [90m 1[39m dose      0.5    1      1-0.5           9.13       6.36    11.9   3.55[90me[39m[31m-10[39m
#> [90m 2[39m dose      0.5    2      2-0.5          15.5       12.7     18.3   4.38[90me[39m[31m-13[39m
#> [90m 3[39m dose      1      2      2-1             6.36       3.60     9.13  2.71[90me[39m[31m- 6[39m
#> [90m 4[39m supp      OJ     VC     VC-OJ          -[31m3[39m[31m.[39m[31m7[39m       -[31m5[39m[31m.[39m[31m58[39m    -[31m1[39m[31m.[39m[31m82[39m  2.31[90me[39m[31m- 4[39m
#> [90m 5[39m dose:supp 0.5:OJ 1:OJ   1:OJ-0.5:OJ     9.47       4.67    14.3   4.61[90me[39m[31m- 6[39m
#> [90m 6[39m dose:supp 0.5:OJ 2:OJ   2:OJ-0.5:OJ    12.8        8.03    17.6   2.13[90me[39m[31m- 9[39m
#> [90m 7[39m dose:supp 0.5:OJ 0.5:VC 0.5:VC-0.5:OJ  -[31m5[39m[31m.[39m[31m25[39m     -[31m10[39m[31m.[39m[31m0[39m     -[31m0[39m[31m.[39m[31m452[39m 2.43[90me[39m[31m- 2[39m
#> [90m 8[39m dose:supp 0.5:OJ 1:VC   1:VC-0.5:OJ     3.54      -[31m1[39m[31m.[39m[31m26[39m     8.34  2.64[90me[39m[31m- 1[39m
#> [90m 9[39m dose:supp 0.5:OJ 2:VC   2:VC-0.5:OJ    12.9        8.11    17.7   1.77[90me[39m[31m- 9[39m
#> [90m10[39m dose:supp 1:OJ   2:OJ   2:OJ-1:OJ       3.36      -[31m1[39m[31m.[39m[31m44[39m     8.16  3.19[90me[39m[31m- 1[39m
#> [90m11[39m dose:supp 1:OJ   0.5:VC 0.5:VC-1:OJ   -[31m14[39m[31m.[39m[31m7[39m      -[31m19[39m[31m.[39m[31m5[39m     -[31m9[39m[31m.[39m[31m92[39m  2.99[90me[39m[31m-11[39m
#> [90m12[39m dose:supp 1:OJ   1:VC   1:VC-1:OJ      -[31m5[39m[31m.[39m[31m93[39m     -[31m10[39m[31m.[39m[31m7[39m     -[31m1[39m[31m.[39m[31m13[39m  7.39[90me[39m[31m- 3[39m
#> [90m13[39m dose:supp 1:OJ   2:VC   2:VC-1:OJ       3.44      -[31m1[39m[31m.[39m[31m36[39m     8.24  2.94[90me[39m[31m- 1[39m
#> [90m14[39m dose:supp 2:OJ   0.5:VC 0.5:VC-2:OJ   -[31m18[39m[31m.[39m[31m1[39m      -[31m22[39m[31m.[39m[31m9[39m    -[31m13[39m[31m.[39m[31m3[39m   4.86[90me[39m[31m-13[39m
#> [90m15[39m dose:supp 2:OJ   1:VC   1:VC-2:OJ      -[31m9[39m[31m.[39m[31m29[39m     -[31m14[39m[31m.[39m[31m1[39m     -[31m4[39m[31m.[39m[31m49[39m  6.91[90me[39m[31m- 6[39m
#> [90m16[39m dose:supp 2:OJ   2:VC   2:VC-2:OJ       0.080[4m0[24m    -[31m4[39m[31m.[39m[31m72[39m     4.88  1.00[90me[39m+ 0
#> [90m17[39m dose:supp 0.5:VC 1:VC   1:VC-0.5:VC     8.79       3.99    13.6   2.10[90me[39m[31m- 5[39m
#> [90m18[39m dose:supp 0.5:VC 2:VC   2:VC-0.5:VC    18.2       13.4     23.0   4.82[90me[39m[31m-13[39m
#> [90m19[39m dose:supp 1:VC   2:VC   2:VC-1:VC       9.37       4.57    14.2   5.77[90me[39m[31m- 6[39m

# Tukey HSD from lm() results
lm(len ~ dose, data = df) %>% tukey_hsd()

#> [90m# A tibble: 3 x 8[39m
#>   term  group1 group2 comparison estimate conf.low conf.high    p.adj
#>   [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m         [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m
#> [90m1[39m dose  0.5    1      1-0.5          9.13     5.90     12.4  2.00[90me[39m[31m- 8[39m
#> [90m2[39m dose  0.5    2      2-0.5         15.5     12.3      18.7  1.12[90me[39m[31m-11[39m
#> [90m3[39m dose  1      2      2-1            6.36     3.14      9.59 4.25[90me[39m[31m- 5[39m

# Tukey HSD from data frame and formula
tukey_hsd(df, len ~ dose)

#> [90m# A tibble: 3 x 8[39m
#>   term  group1 group2 comparison estimate conf.low conf.high    p.adj
#>   [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m         [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m
#> [90m1[39m dose  0.5    1      1-0.5          9.13     5.90     12.4  2.00[90me[39m[31m- 8[39m
#> [90m2[39m dose  0.5    2      2-0.5         15.5     12.3      18.7  1.12[90me[39m[31m-11[39m
#> [90m3[39m dose  1      2      2-1            6.36     3.14      9.59 4.25[90me[39m[31m- 5[39m

# Tukey HSD using grouped data
df %>%
  group_by(supp) %>%
  tukey_hsd(len ~ dose)

#> [90m# A tibble: 6 x 9[39m
#>   supp  term  group1 group2 comparison estimate conf.low conf.high    p.adj
#>   [3m[90m<fct>[39m[23m [3m[90m<chr>[39m[23m [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m  [3m[90m<chr>[39m[23m         [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m     [3m[90m<dbl>[39m[23m    [3m[90m<dbl>[39m[23m
#> [90m1[39m OJ    dose  0.5    1      1-0.5          9.47    5.31      13.6  1.58[90me[39m[31m- 5[39m
#> [90m2[39m OJ    dose  0.5    2      2-0.5         12.8     8.67      17.0  9.39[90me[39m[31m- 8[39m
#> [90m3[39m OJ    dose  1      2      2-1            3.36   -[31m0[39m[31m.[39m[31m800[39m      7.52 1.31[90me[39m[31m- 1[39m
#> [90m4[39m VC    dose  0.5    1      1-0.5          8.79    4.90      12.7  1.75[90me[39m[31m- 5[39m
#> [90m5[39m VC    dose  0.5    2      2-0.5         18.2    14.3       22.0  1.66[90me[39m[31m-11[39m
#> [90m6[39m VC    dose  1      2      2-1            9.37    5.48      13.3  6.61[90me[39m[31m- 6[39m