Tukey Honest Significant Differences

Provides a pipe-friendly framework to performs Tukey post-hoc tests. Wrapper around the function TukeyHSD().

tukey_hsd(x, ...)

tukey_hsd2(data, formula, ...)

Arguments

x

A fitted model object, usually an aov fit.

...

other arguments passed to the function TukeyHSD().

data

a data.frame containing the variables in the formula.

formula

a formula of the form x ~ group where x is a numeric variable giving the data values and group is a factor with one or multiple levels giving the corresponding groups. For example, formula = TP53 ~ cancer_group.

Value

a tibble data frame containing the results of the different comparisons.

Functions

tukey_hsd: performs tukey post-hoc test from ANOVA results, usually an aov fit.

tukey_hsd2: performs tukey post-hoc tests using data and formula as inputs. ANOVA will be automatically performed using the function aov()

Examples

# Data preparation
df <- ToothGrowth
df$dose <- as.factor(df$dose)
# Anova
res.aov <- aov(len ~ dose, data = df)
# Tukey hsd
tukey_hsd(res.aov)

#> # A tibble: 3 x 8
#>    term group1 group2 comparison estimate  conf.low conf.high   p.adj
#>   <chr>  <chr>  <chr>      <chr>    <dbl>     <dbl>     <dbl>   <dbl>
#> 1  dose    0.5      1      1-0.5    9.130  5.901805 12.358195 2.0e-08
#> 2  dose    0.5      2      2-0.5   15.495 12.266805 18.723195 1.1e-11
#> 3  dose      1      2        2-1    6.365  3.136805  9.593195 4.2e-05