Wilcoxon Tests

Provides a pipe-friendly framework to performs one and two sample Wilcoxon tests.

wilcox_test(data, formula, comparisons = NULL, ref.group = NULL,
  p.adjust.method = "holm", paired = FALSE, exact = FALSE,
  alternative = "two.sided", mu = 0, conf.level = 0.95)

one_sample_wilcox_test(data, formula, mu = 0, exact = FALSE, ...)

two_sample_wilcox_test(data, formula, paired = FALSE, exact = FALSE, ...)

pairwise_wilcox_test(data, formula, comparisons = NULL, ref.group = NULL,
  p.adjust.method = "holm", ...)

one_vs_all_wilcox_test(data, formula, p.adjust.method = "holm", ...)

Arguments

data

a data.frame containing the variables in the formula.

formula

a formula of the form x ~ group where x is a numeric variable giving the data values and group is a factor with one or multiple levels giving the corresponding groups. For example, formula = TP53 ~ cancer_group.

comparisons

A list of length-2 vectors specifying the groups of interest to be compared. For example to compare groups "A" vs "B" and "B" vs "C", the argument is as follow: comparisons = list(c("A", "B"), c("B", "C"))

ref.group

a character string specifying the reference group. If specified, for a given grouping variable, each of the group levels will be compared to the reference group (i.e. control group).

p.adjust.method

method to adjust p values for multiple comparisons. Used when pairwise comparisons are performed. Allowed values include "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". If you don't want to adjust the p value (not recommended), use p.adjust.method = "none".

paired

a logical indicating whether you want a paired test.

exact

a logical indicating whether an exact p-value should be computed.

alternative

a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.

mu

a number specifying an optional parameter used to form the null hypothesis. See ‘Details’.

conf.level

confidence level of the interval.

...

other arguments to be passed to the function wilcox.test.

Value

return a data frame with the following columns:

.y.: the y variable used in the test.

group1,group2: the compared groups in the pairwise tests.

statistic: Test statistic used to compute the p-value.

p: p-value.

p.adj: the adjusted p-value.

method: the statistical test used to compare groups.

p.signif, p.adj.signif: the significance level of p-values and adjusted p-values, respectively.

Details

- pairwise_wilcox_test() applies the standard two sample Wilcoxon test to all possible pairs of groups. This method calls the wilcox.test(), so extra arguments are accepted.

- If a list of comparisons is specified, the result of the pairwise tests is filtered to keep only the comparisons of interest.The p-value is adjusted after filtering.

Functions

one_sample_wilcox_test: performs one-sample Wilcoxon test.

two_sample_wilcox_test: performs two sample Wilcoxon test.

pairwise_wilcox_test: performs pairwise two sample Wilcoxon test.

one_vs_all_wilcox_test: performs pairwise two sample Wilcoxon test comparing each grouping variable levels against all (i.e. basemean)

Examples

# Load data
#:::::::::::::::::::::::::::::::::::::::
data("ToothGrowth")
df <- ToothGrowth

# One-sample test
#:::::::::::::::::::::::::::::::::::::::::
df %>% wilcox_test(len ~ 1, mu = 0)

#> # A tibble: 1 x 6
#>     .y. group1     group2 statistic       p   method
#>   <chr>  <chr>      <chr>     <dbl>   <dbl>    <chr>
#> 1   len      1 null model      1830 1.7e-11 Wilcoxon

# Two-samples unpaired test
#:::::::::::::::::::::::::::::::::::::::::
df %>% wilcox_test(len ~ supp)

#> # A tibble: 1 x 6
#>     .y. group1 group2 statistic     p   method
#>   <chr>  <chr>  <chr>     <dbl> <dbl>    <chr>
#> 1   len     OJ     VC     575.5 0.064 Wilcoxon

# Two-samples paired test
#:::::::::::::::::::::::::::::::::::::::::
df %>% wilcox_test (len ~ supp, paired = TRUE)

#> # A tibble: 1 x 6
#>     .y. group1 group2 statistic      p   method
#>   <chr>  <chr>  <chr>     <dbl>  <dbl>    <chr>
#> 1   len     OJ     VC       350 0.0043 Wilcoxon

# Compare supp levels after grouping the data by "dose"
#::::::::::::::::::::::::::::::::::::::::
df %>%
  group_by(dose) %>%
  do(wilcox_test(data =., len ~ supp)) %>%
  ungroup() %>%
  adjust_pvalue() %>%
  add_significance("p.adj")

#> # A tibble: 3 x 9
#>    dose   .y. group1 group2 statistic     p   method p.adj p.adj.signif
#>   <dbl> <chr>  <chr>  <chr>     <dbl> <dbl>    <chr> <dbl>        <chr>
#> 1   0.5   len     OJ     VC      80.5 0.023 Wilcoxon 0.046            *
#> 2   1.0   len     OJ     VC      88.5 0.004 Wilcoxon 0.012            *
#> 3   2.0   len     OJ     VC      49.5 1.000 Wilcoxon 1.000           ns

# pairwise comparisons
#::::::::::::::::::::::::::::::::::::::::
# As dose contains more thant two levels ==>
# pairwise test is automatically performed.
df %>% wilcox_test(len ~ dose)

#> # A tibble: 3 x 9
#>     .y. group1 group2 statistic       p   method   p.adj p.signif p.adj.signif
#>   <chr>  <chr>  <chr>     <dbl>   <dbl>    <chr>   <dbl>    <chr>        <chr>
#> 1   len    0.5      1      33.5 7.0e-06 Wilcoxon 1.4e-05     ****         ****
#> 2   len    0.5      2       1.5 8.4e-08 Wilcoxon 2.5e-07     ****         ****
#> 3   len      1      2      61.0 1.8e-04 Wilcoxon 1.8e-04      ***          ***

# Comparison against reference group
#::::::::::::::::::::::::::::::::::::::::
# each level is compared to the ref group
df %>% wilcox_test(len ~ dose, ref.group = "0.5")

#> # A tibble: 2 x 9
#>     .y. group1 group2 statistic       p   method   p.adj p.signif p.adj.signif
#>   <chr>  <chr>  <chr>     <dbl>   <dbl>    <chr>   <dbl>    <chr>        <chr>
#> 1   len    0.5      1      33.5 7.0e-06 Wilcoxon 7.0e-06     ****         ****
#> 2   len    0.5      2       1.5 8.4e-08 Wilcoxon 1.7e-07     ****         ****

# Comparison against all
#::::::::::::::::::::::::::::::::::::::::
df %>% wilcox_test(len ~ dose, ref.group = "all")

#> # A tibble: 3 x 9
#>     .y. group1 group2 statistic       p   method   p.adj p.signif p.adj.signif
#>   <chr>  <chr>  <chr>     <dbl>   <dbl>    <chr>   <dbl>    <chr>        <chr>
#> 1   len    all    0.5     965.0 5.1e-05 Wilcoxon 0.00015     ****          ***
#> 2   len    all      1     572.5 7.6e-01 Wilcoxon 0.76000       ns           ns
#> 3   len    all      2     262.5 1.8e-04 Wilcoxon 0.00036      ***          ***